kitt/requirements.txt
# Core Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# MCP Server
mcp==1.21.2
fastmcp==2.13.1

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.25
neo4j==5.14.1  # Neo4j graph database driver

# Streaming
kafka-python==2.0.2

# LLM
anthropic==0.74.0

# Packing Engine (will install separately)
# git+https://github.com/zgtcktom/DeepPack3D.git
tensorflow==2.10.0
numpy==1.24.3
matplotlib==3.9.0

# ML/Prediction
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# External APIs
httpx==0.26.0
aiohttp==3.9.1

# Caching
redis==5.0.1
aiocache==0.12.2

# Utilities
python-dotenv==1.0.0
